+++ src/pyatmo/__init__.py
@@ -1,8 +1,17 @@
 """Expose submodules."""
+
 from pyatmo import const, modules
 from pyatmo.account import AsyncAccount
 from pyatmo.auth import AbstractAsyncAuth
-from pyatmo.exceptions import ApiError, InvalidHome, InvalidRoom, NoDevice, NoSchedule
+from pyatmo.exceptions import (
+    ApiError,
+    ApiErrorThrottling,
+    ApiHomeReachabilityError,
+    InvalidHome,
+    InvalidRoom,
+    NoDevice,
+    NoSchedule,
+)
 from pyatmo.home import Home
 from pyatmo.modules import Module
 from pyatmo.modules.device_types import DeviceType
@@ -11,6 +20,8 @@ from pyatmo.room import Room
 __all__ = [
     "AbstractAsyncAuth",
     "ApiError",
+    "ApiErrorThrottling",
+    "ApiHomeReachabilityError",
     "AsyncAccount",
     "InvalidHome",
     "InvalidRoom",