+++ src/pyatmo/const.py
@@ -1,4 +1,5 @@
 """Common constants."""
+
 from __future__ import annotations
 
 from typing import Any
@@ -87,7 +88,6 @@ FROSTGUARD = "hg"
 SCHEDULES = "schedules"
 EVENTS = "events"
 
-
 STATION_TEMPERATURE_TYPE = "temperature"
 STATION_PRESSURE_TYPE = "pressure"
 STATION_HUMIDITY_TYPE = "humidity"
@@ -101,3 +101,8 @@ ACCESSORY_WIND_ANGLE_TYPE = "wind_angle"
 ACCESSORY_WIND_TIME_TYPE = "wind_timeutc"
 ACCESSORY_GUST_STRENGTH_TYPE = "gust_strength"
 ACCESSORY_GUST_ANGLE_TYPE = "gust_angle"
+
+# 2 days of dynamic historical data stored
+MAX_HISTORY_TIME_FRAME = 24 * 2 * 3600
+
+UNKNOWN = "unknown"