+++ src/pyatmo/modules/module.py
@@ -1,7 +1,8 @@
 """Module to represent a Netatmo module."""
+
 from __future__ import annotations
 
-from datetime import datetime, timezone
+from datetime import datetime, timedelta, timezone
 from enum import Enum
 import logging
 from typing import TYPE_CHECKING, Any
@@ -16,6 +17,9 @@ from pyatmo.modules.device_types import DEVICE_CATEGORY_MAP, DeviceCategory, Dev
 if TYPE_CHECKING:
     from pyatmo.event import Event
     from pyatmo.home import Home
+
+from operator import itemgetter
+from time import time
 
 LOG = logging.getLogger(__name__)
 
@@ -40,6 +44,8 @@ ATTRIBUTE_FILTER = {
     "device_category",
     "device_type",
     "features",
+    "history_features",
+    "history_features_values",
 }
 
 
@@ -222,6 +228,16 @@ class BoilerMixin(EntityBase):
         self.boiler_status: bool | None = None
 
 
+class CoolerMixin(EntityBase):
+    """Mixin for cooler data."""
+
+    def __init__(self, home: Home, module: ModuleT):
+        """Initialize cooler mixin."""
+
+        super().__init__(home, module)  # type: ignore # mypy issue 4335
+        self.cooler_status: bool | None = None
+
+
 class BatteryMixin(EntityBase):
     """Mixin for battery data."""
 
@@ -288,16 +304,6 @@ class ApplianceTypeMixin(EntityBase):
         self.appliance_type: str | None = None
 
 
-class EnergyMixin(EntityBase):
-    """Mixin for energy data."""
-
-    def __init__(self, home: Home, module: ModuleT):
-        """Initialize energy mixin."""
-
-        super().__init__(home, module)  # type: ignore # mypy issue 4335
-        self.sum_energy_elec: int | None = None
-
-
 class PowerMixin(EntityBase):
     """Mixin for power data."""
 
@@ -306,6 +312,7 @@ class PowerMixin(EntityBase):
 
         super().__init__(home, module)  # type: ignore # mypy issue 4335
         self.power: int | None = None
+        self.history_features.add("power")
 
 
 class EventMixin(EntityBase):
@@ -402,21 +409,32 @@ class FanSpeedMixin(EntityBase):
 class ShutterMixin(EntityBase):
     """Mixin for shutter data."""
 
+    __open_position = 100
+    __close_position = 0
+    __stop_position = -1
+    __preferred_position = -2
+
     def __init__(self, home: Home, module: ModuleT):
         """Initialize shutter mixin."""
 
         super().__init__(home, module)  # type: ignore # mypy issue 4335
         self.current_position: int | None = None
         self.target_position: int | None = None
+        self.target_position__step: int | None = None
 
     async def async_set_target_position(self, target_position: int) -> bool:
         """Set shutter to target position."""
+
+        # in case of a too low value, we default to stop and not the preferred position
+        # We check against __preferred_position that is the lower known value
+        if target_position < self.__preferred_position:
+            target_position = self.__stop_position
 
         json_roller_shutter = {
             "modules": [
                 {
                     "id": self.entity_id,
-                    "target_position": max(min(100, target_position), -1),
+                    "target_position": min(self.__open_position, target_position),
                     "bridge": self.bridge,
                 },
             ],
@@ -426,17 +444,22 @@ class ShutterMixin(EntityBase):
     async def async_open(self) -> bool:
         """Open shutter."""
 
-        return await self.async_set_target_position(100)
+        return await self.async_set_target_position(self.__open_position)
 
     async def async_close(self) -> bool:
         """Close shutter."""
 
-        return await self.async_set_target_position(0)
+        return await self.async_set_target_position(self.__close_position)
 
     async def async_stop(self) -> bool:
         """Stop shutter."""
 
-        return await self.async_set_target_position(-1)
+        return await self.async_set_target_position(self.__stop_position)
+
+    async def async_move_to_preferred_position(self) -> bool:
+        """Move shutter to preferred position."""
+
+        return await self.async_set_target_position(self.__preferred_position)
 
 
 class CameraMixin(EntityBase):
@@ -601,18 +624,64 @@ class MeasureType(Enum):
     BOILEROFF = "boileroff"
     SUM_BOILER_ON = "sum_boiler_on"
     SUM_BOILER_OFF = "sum_boiler_off"
-    SUM_ENERGY_ELEC = "sum_energy_elec"
-    SUM_ENERGY_ELEC_BASIC = "sum_energy_elec$0"
-    SUM_ENERGY_ELEC_PEAK = "sum_energy_elec$1"
-    SUM_ENERGY_ELEC_OFF_PEAK = "sum_energy_elec$2"
-    SUM_ENERGY_PRICE = "sum_energy_price"
-    SUM_ENERGY_PRICE_BASIC = "sum_energy_price$0"
-    SUM_ENERGY_PRICE_PEAK = "sum_energy_price$1"
-    SUM_ENERGY_PRICE_OFF_PEAK = "sum_energy_price$2"
+    SUM_ENERGY_ELEC = "sum_energy_buy_from_grid"
+    SUM_ENERGY_ELEC_BASIC = "sum_energy_buy_from_grid$0"
+    SUM_ENERGY_ELEC_PEAK = "sum_energy_buy_from_grid$1"
+    SUM_ENERGY_ELEC_OFF_PEAK = "sum_energy_buy_from_grid$2"
+    SUM_ENERGY_PRICE = "sum_energy_buy_from_grid_price"
+    SUM_ENERGY_PRICE_BASIC = "sum_energy_buy_from_grid_price$0"
+    SUM_ENERGY_PRICE_PEAK = "sum_energy_buy_from_grid_price$1"
+    SUM_ENERGY_PRICE_OFF_PEAK = "sum_energy_buy_from_grid_price$2"
+    SUM_ENERGY_ELEC_LEGACY = "sum_energy_elec"
+    SUM_ENERGY_ELEC_BASIC_LEGACY = "sum_energy_elec$0"
+    SUM_ENERGY_ELEC_PEAK_LEGACY = "sum_energy_elec$1"
+    SUM_ENERGY_ELEC_OFF_PEAK_LEGACY = "sum_energy_elec$2"
 
 
-class HistoryMixin(EntityBase):
-    """Mixin for history data."""
+MEASURE_INTERVAL_TO_SECONDS = {
+    MeasureInterval.HALF_HOUR: 1800,
+    MeasureInterval.HOUR: 3600,
+    MeasureInterval.THREE_HOURS: 10800,
+    MeasureInterval.DAY: 86400,
+    MeasureInterval.WEEK: 604800,
+    MeasureInterval.MONTH: 2592000,
+}
+
+ENERGY_FILTERS = f"{MeasureType.SUM_ENERGY_ELEC.value},{MeasureType.SUM_ENERGY_ELEC_BASIC.value},{MeasureType.SUM_ENERGY_ELEC_PEAK.value},{MeasureType.SUM_ENERGY_ELEC_OFF_PEAK.value}"
+ENERGY_FILTERS_LEGACY = f"{MeasureType.SUM_ENERGY_ELEC_LEGACY.value},{MeasureType.SUM_ENERGY_ELEC_BASIC_LEGACY.value},{MeasureType.SUM_ENERGY_ELEC_PEAK_LEGACY.value},{MeasureType.SUM_ENERGY_ELEC_OFF_PEAK_LEGACY.value}"
+ENERGY_FILTERS_MODES = ["generic", "basic", "peak", "off_peak"]
+
+
+def compute_riemann_sum(
+    power_data: list[tuple[int, float]], conservative: bool = False
+):
+    """Compute energy from power with a rieman sum."""
+
+    delta_energy = 0
+    if power_data and len(power_data) > 1:
+        # compute a rieman sum, as best as possible , trapezoidal, taking pessimistic asumption
+        # as we don't want to artifically go up the previous one
+        # (except in rare exceptions like reset, 0 , etc)
+
+        for i in range(len(power_data) - 1):
+            dt_h = float(power_data[i + 1][0] - power_data[i][0]) / 3600.0
+
+            if conservative:
+                d_p_w = 0
+            else:
+                d_p_w = abs(float(power_data[i + 1][1] - power_data[i][1]))
+
+            d_nrj_wh = dt_h * (
+                min(power_data[i + 1][1], power_data[i][1]) + 0.5 * d_p_w
+            )
+
+            delta_energy += d_nrj_wh
+
+    return delta_energy
+
+
+class EnergyHistoryMixin(EntityBase):
+    """Mixin for Energy history data."""
 
     def __init__(self, home: Home, module: ModuleT):
         """Initialize history mixin."""
@@ -620,26 +689,300 @@ class HistoryMixin(EntityBase):
         super().__init__(home, module)  # type: ignore # mypy issue 4335
         self.historical_data: list[dict[str, Any]] | None = None
         self.start_time: int | None = None
+        self.end_time: int | None = None
         self.interval: MeasureInterval | None = None
+        self.sum_energy_elec: int | None = None
+        self.sum_energy_elec_peak: int | None = None
+        self.sum_energy_elec_off_peak: int | None = None
+        self._anchor_for_power_adjustment: int | None = None
+        self.in_reset: bool | False = False
+
+    def reset_measures(self, start_power_time, in_reset=True):
+        """Reset energy measures."""
+        self.in_reset = in_reset
+        self.historical_data = []
+        if start_power_time is None:
+            self._anchor_for_power_adjustment = start_power_time
+        else:
+            self._anchor_for_power_adjustment = int(start_power_time.timestamp())
+        self.sum_energy_elec = 0
+        self.sum_energy_elec_peak = 0
+        self.sum_energy_elec_off_peak = 0
+
+    def get_sum_energy_elec_power_adapted(
+        self, to_ts: int | float | None = None, conservative: bool = False
+    ):
+        """Compute proper energy value with adaptation from power."""
+        v = self.sum_energy_elec
+
+        if v is None:
+            return None, 0
+
+        delta_energy = 0
+
+        if not self.in_reset:
+            if to_ts is None:
+                to_ts = int(time())
+
+            from_ts = self._anchor_for_power_adjustment
+
+            if (
+                from_ts is not None
+                and from_ts < to_ts
+                and isinstance(self, PowerMixin)
+                and isinstance(self, NetatmoBase)
+            ):
+                power_data = self.get_history_data(
+                    "power", from_ts=from_ts, to_ts=to_ts
+                )
+                if isinstance(
+                    self, EnergyHistoryMixin
+                ):  # well to please the linter....
+                    delta_energy = compute_riemann_sum(power_data, conservative)
+
+        return v, delta_energy
+
+    def _log_energy_error(self, start_time, end_time, msg=None, body=None):
+        if body is None:
+            body = "NO BODY"
+        LOG.debug(
+            "ENERGY collection error %s %s %s %s %s %s %s",
+            msg,
+            self.name,
+            datetime.fromtimestamp(start_time),
+            datetime.fromtimestamp(end_time),
+            start_time,
+            end_time,
+            body,
+        )
 
     async def async_update_measures(
         self,
         start_time: int | None = None,
+        end_time: int | None = None,
         interval: MeasureInterval = MeasureInterval.HOUR,
         days: int = 7,
     ) -> None:
         """Update historical data."""
 
-        end_time = int(datetime.now().timestamp())
-        if start_time is None:
-            start_time = end_time - days * 24 * 60 * 60
+        if end_time is None:
+            end_time = int(datetime.now().timestamp())
 
-        data_point = MeasureType.SUM_ENERGY_ELEC_BASIC.value
+        if start_time is None:
+            end = datetime.fromtimestamp(end_time)
+            start_time = end - timedelta(days=days)
+            start_time = int(start_time.timestamp())
+
+        prev_start_time = self.start_time
+        prev_end_time = self.end_time
+
+        self.start_time = start_time
+        self.end_time = end_time
+
+        # the legrand/netatmo handling of start and endtime is very peculiar
+        # for 30mn/1h/3h intervals : in fact the starts is asked_start + intervals/2 !
+        # => so shift of 15mn, 30mn and 1h30
+        # for 1day : start is ALWAYS 12am (half day) of the first day of the range
+        # for 1week : it will be half week ALWAYS, ie on a thursday at 12am (half day)
+        # in fact in the case for all intervals the reported dates are "the middle" of the ranges
+
+        delta_range = MEASURE_INTERVAL_TO_SECONDS.get(interval, 0) // 2
+
+        filters, raw_data = await self._energy_API_calls(start_time, end_time, interval)
+
+        hist_good_vals = await self._get_aligned_energy_values_and_mode(
+            start_time, end_time, delta_range, raw_data
+        )
+
+        self.historical_data = []
+        prev_sum_energy_elec = self.sum_energy_elec
+        self.sum_energy_elec = 0
+        self.sum_energy_elec_peak = 0
+        self.sum_energy_elec_off_peak = 0
+
+        # no data at all: we know nothing for the end: best guess, it is the start
+        self._anchor_for_power_adjustment = start_time
+
+        self.in_reset = False
+
+        if len(hist_good_vals) == 0:
+            # nothing has been updated or changed it can nearly be seen as an error, but the api is answering correctly
+            # so we probably have to reset to 0 anyway as it means there were no exisitng
+            # historical data for this time range
+
+            LOG.debug(
+                "NO VALUES energy update %s from: %s to %s,  prev_sum=%s",
+                self.name,
+                datetime.fromtimestamp(start_time),
+                datetime.fromtimestamp(end_time),
+                prev_sum_energy_elec if prev_sum_energy_elec is not None else "NOTHING",
+            )
+        else:
+            await self._prepare_exported_historical_data(
+                start_time,
+                end_time,
+                delta_range,
+                hist_good_vals,
+                prev_end_time,
+                prev_start_time,
+                prev_sum_energy_elec,
+            )
+
+    async def _prepare_exported_historical_data(
+        self,
+        start_time,
+        end_time,
+        delta_range,
+        hist_good_vals,
+        prev_end_time,
+        prev_start_time,
+        prev_sum_energy_elec,
+    ):
+        computed_start = 0
+        computed_end = 0
+        computed_end_for_calculus = 0
+        for cur_start_time, val, vals in hist_good_vals:
+            self.sum_energy_elec += val
+
+            modes = []
+            val_modes = []
+
+            for i, v in enumerate(vals):
+                if v is not None:
+                    modes.append(ENERGY_FILTERS_MODES[i])
+                    val_modes.append(v)
+                    if ENERGY_FILTERS_MODES[i] == "off_peak":
+                        self.sum_energy_elec_off_peak += v
+                    elif ENERGY_FILTERS_MODES[i] == "peak":
+                        self.sum_energy_elec_peak += v
+
+            c_start = cur_start_time
+            c_end = cur_start_time + 2 * delta_range
+
+            if computed_start == 0:
+                computed_start = c_start
+            computed_end = c_end
+
+            # - delta_range not sure, revert ... it seems the energy value effectively stops at those mid values
+            computed_end_for_calculus = c_end  # - delta_range
+
+            start_time_string = f"{datetime.fromtimestamp(c_start + 1, tz=timezone.utc).isoformat().split('+')[0]}Z"
+            end_time_string = f"{datetime.fromtimestamp(c_end, tz=timezone.utc).isoformat().split('+')[0]}Z"
+            self.historical_data.append(
+                {
+                    "duration": (2 * delta_range) // 60,
+                    "startTime": start_time_string,
+                    "endTime": end_time_string,
+                    "Wh": val,
+                    "energyMode": modes,
+                    "WhPerModes": val_modes,
+                    "startTimeUnix": c_start,
+                    "endTimeUnix": c_end,
+                },
+            )
+        if (
+            prev_sum_energy_elec is not None
+            and prev_sum_energy_elec > self.sum_energy_elec
+        ):
+            msg = (
+                "ENERGY GOING DOWN %s from: %s to %s "
+                "computed_start: %s, computed_end: %s, "
+                "sum=%f prev_sum=%f prev_start: %s, prev_end %s"
+            )
+            LOG.debug(
+                msg,
+                self.name,
+                datetime.fromtimestamp(start_time),
+                datetime.fromtimestamp(end_time),
+                datetime.fromtimestamp(computed_start),
+                datetime.fromtimestamp(computed_end),
+                self.sum_energy_elec,
+                prev_sum_energy_elec,
+                datetime.fromtimestamp(prev_start_time),
+                datetime.fromtimestamp(prev_end_time),
+            )
+        else:
+            msg = (
+                "Success in energy update %s from: %s to %s "
+                "computed_start: %s, computed_end: %s , sum=%s prev_sum=%s"
+            )
+            LOG.debug(
+                msg,
+                self.name,
+                datetime.fromtimestamp(start_time),
+                datetime.fromtimestamp(end_time),
+                datetime.fromtimestamp(computed_start),
+                datetime.fromtimestamp(computed_end),
+                self.sum_energy_elec,
+                prev_sum_energy_elec if prev_sum_energy_elec is not None else "NOTHING",
+            )
+
+        self._anchor_for_power_adjustment = computed_end_for_calculus
+
+    async def _get_aligned_energy_values_and_mode(
+        self, start_time, end_time, delta_range, raw_data
+    ):
+        hist_good_vals = []
+        values_lots = raw_data
+
+        for values_lot in values_lots:
+            try:
+                start_lot_time = int(values_lot["beg_time"])
+            except Exception:
+                self._log_energy_error(
+                    start_time,
+                    end_time,
+                    msg="beg_time missing",
+                    body=values_lots,
+                )
+                raise ApiError(
+                    f"Energy badly formed resp beg_time missing: {values_lots} - "
+                    f"module: {self.name}"
+                ) from None
+
+            interval_sec = values_lot.get("step_time")
+            if interval_sec is None:
+                if len(values_lot.get("value", [])) > 1:
+                    self._log_energy_error(
+                        start_time,
+                        end_time,
+                        msg="step_time missing",
+                        body=values_lots,
+                    )
+                interval_sec = 2 * delta_range
+            else:
+                interval_sec = int(interval_sec)
+
+            # align the start on the begining of the segment
+            cur_start_time = start_lot_time - interval_sec // 2
+            for val_arr in values_lot.get("value", []):
+                vals = []
+                val = 0
+                for v in val_arr:
+                    if v is not None:
+                        v = int(v)
+                        val += v
+                        vals.append(v)
+                    else:
+                        vals.append(None)
+
+                hist_good_vals.append((cur_start_time, val, vals))
+                cur_start_time = cur_start_time + interval_sec
+
+        hist_good_vals = sorted(hist_good_vals, key=itemgetter(0))
+        return hist_good_vals
+
+    def _get_energy_filers(self):
+        return ENERGY_FILTERS
+
+    async def _energy_API_calls(self, start_time, end_time, interval):
+        filters = self._get_energy_filers()
+
         params = {
             "device_id": self.bridge,
             "module_id": self.entity_id,
             "scale": interval.value,
-            "type": data_point,
+            "type": filters,
             "date_begin": start_time,
             "date_end": end_time,
         }
@@ -648,27 +991,30 @@ class HistoryMixin(EntityBase):
             endpoint=GETMEASURE_ENDPOINT,
             params=params,
         )
-        raw_data = await resp.json()
 
-        data = raw_data["body"][0]
-        interval_sec = int(data["step_time"])
-        interval_min = interval_sec // 60
+        rw_dt_f = await resp.json()
+        rw_dt = rw_dt_f.get("body")
 
-        self.historical_data = []
-        self.start_time = int(data["beg_time"])
-        start_time = self.start_time
-        for value in data["value"]:
-            end_time = start_time + interval_sec
-            self.historical_data.append(
-                {
-                    "duration": interval_min,
-                    "startTime": f"{datetime.fromtimestamp(start_time + 1, tz=timezone.utc).isoformat().split('+')[0]}Z",
-                    "endTime": f"{datetime.fromtimestamp(end_time, tz=timezone.utc).isoformat().split('+')[0]}Z",
-                    "Wh": value[0],
-                },
+        if rw_dt is None:
+            self._log_energy_error(
+                start_time, end_time, msg=f"direct from {filters}", body=rw_dt_f
+            )
+            raise ApiError(
+                f"Energy badly formed resp: {rw_dt_f} - "
+                f"module: {self.name} - "
+                f"when accessing '{filters}'"
             )
 
-            start_time = end_time
+        raw_data = rw_dt
+
+        return filters, raw_data
+
+
+class EnergyHistoryLegacyMixin(EnergyHistoryMixin):
+    """Mixin for Energy history data, Using legacy APis (used for NLE)."""
+
+    def _get_energy_filers(self):
+        return ENERGY_FILTERS_LEGACY
 
 
 class Module(NetatmoBase):
@@ -702,6 +1048,15 @@ class Module(NetatmoBase):
 
         self.update_topology(raw_data)
         self.update_features()
+
+        # If we have an NLE as a bridge all its bridged modules will have to be reachable
+        if self.device_type == DeviceType.NLE:
+            # if there is a bridge it means it is a leaf
+            if self.bridge:
+                self.reachable = True
+            elif self.modules:
+                # this NLE is a bridge itself : make it not available
+                self.reachable = False
 
         if not self.reachable and self.modules:
             # Update bridged modules and associated rooms
@@ -742,7 +1097,7 @@ class Camera(
         await self.async_update_camera_urls()
 
 
-class Switch(FirmwareMixin, PowerMixin, SwitchMixin, Module):
+class Switch(FirmwareMixin, EnergyHistoryMixin, PowerMixin, SwitchMixin, Module):
     """Class to represent a Netatmo switch."""
 
     ...