+++ src/pyatmo/person.py
@@ -1,4 +1,5 @@
 """Module to represent a Netatmo person."""
+
 from __future__ import annotations
 
 from dataclasses import dataclass
@@ -20,6 +21,8 @@ class Person(NetatmoBase):
 
     pseudo: str | None
     url: str | None
+    out_of_sight: bool | None = None
+    last_seen: int | None = None
 
     def __init__(self, home: Home, raw_data: RawData) -> None:
         """Initialize a Netatmo person instance."""
@@ -28,3 +31,8 @@ class Person(NetatmoBase):
         self.home = home
         self.pseudo = raw_data.get("pseudo")
         self.url = raw_data.get("url")
+
+    def update(self, raw_data: RawData) -> None:
+        """Update person data."""
+        self.out_of_sight = raw_data.get("out_of_sight")
+        self.last_seen = raw_data.get("last_seen")