+++ tests/conftest.py
@@ -1,4 +1,5 @@
 """Define shared fixtures."""
+
 # pylint: disable=redefined-outer-name, protected-access
 from contextlib import contextmanager
 from unittest.mock import AsyncMock, patch
@@ -6,7 +7,7 @@ from unittest.mock import AsyncMock, patch
 import pyatmo
 import pytest
 
-from .common import fake_post_request
+from .common import fake_post_request, fake_post_request_multi
 
 
 @contextmanager
@@ -43,3 +44,29 @@ async def async_home(async_account):
     home_id = "91763b24c43d3e344f424e8b"
     await async_account.async_update_status(home_id)
     yield async_account.homes[home_id]
+
+
+@pytest.fixture(scope="function")
+async def async_account_multi(async_auth):
+    """AsyncAccount fixture."""
+    account = pyatmo.AsyncAccount(async_auth)
+
+    with patch(
+        "pyatmo.auth.AbstractAsyncAuth.async_post_api_request",
+        fake_post_request_multi,
+    ), patch(
+        "pyatmo.auth.AbstractAsyncAuth.async_post_request",
+        fake_post_request_multi,
+    ):
+        await account.async_update_topology(
+            disabled_homes_ids=["eeeeeeeeeffffffffffaaaaa"]
+        )
+        yield account
+
+
+@pytest.fixture(scope="function")
+async def async_home_multi(async_account_multi):
+    """AsyncClimate fixture for home_id 91763b24c43d3e344f424e8b."""
+    home_id = "aaaaaaaaaaabbbbbbbbbbccc"
+    await async_account_multi.async_update_status(home_id)
+    yield async_account_multi.homes[home_id]