+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- 
+-
 
 ### Changed
 
@@ -25,11 +25,30 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- 
+-
 
 ### Security
 
--
+
+## [8.1.0]
+
+### Added
+
+- Expose camera person status
+- Add NLE support
+- Add proper energy support
+- Add cooler support
+- Add BNS support
+
+## [8.0.3]
+
+### Added
+
+- Add NLLF centralized ventilation controller
+
+### Fixed
+
+- Add BNSC switch capability
 
 ## [8.0.3]